+++ packages/unified_distributor/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.1
+
+* GitHub publisher supports `repo` argument to replace `repo-owner` and `repo-name` arguments
+* GitHub publisher supports `release-draft` and `release-prerelease` arguments
+* Remove AppCenter publisher
+
 ## 0.2.0
 
 * feat: Support ohos platform.